% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_call_results.R
\name{clean_all_regions}
\alias{clean_all_regions}
\title{clean_all_regions()}
\usage{
clean_all_regions(
  raw,
  year,
  stat_name,
  substat_name,
  parameter,
  region_id,
  page_nr,
  ipp,
  nuts_nr,
  lau_nr,
  parent_chr,
  all_regions
)
}
\arguments{
\item{raw}{description}

\item{year}{description}

\item{stat_name}{description}

\item{substat_name}{description}

\item{parameter}{description}

\item{region_id}{description}

\item{page_nr}{description}

\item{ipp}{description}

\item{nuts_nr}{description}

\item{lau_nr}{description}

\item{parent_chr}{description}

\item{all_regions}{description}
}
\value{
Tidy data frame containing the retrieved results from the API call
}
\description{
Clean retrieved data (if AllRegions endpoint is being called).
}
