% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_it}
\alias{clean_it}
\title{clean_it()}
\usage{
clean_it(.data)
}
\arguments{
\item{.data}{description}
}
\value{
Tidy data frame containing the retrieved results from the API call
}
\description{
Clean retrieved data.
}
