% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_call_results.R
\name{clean_region}
\alias{clean_region}
\title{clean_region()}
\usage{
clean_region(.data)
}
\arguments{
\item{.data}{description}
}
\value{
Tidy data frame containing the retrieved results from the API call
}
\description{
Clean retrieved data (if region id provided).
}
