% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pre_call.R
\name{define_fields}
\alias{define_fields}
\title{define_fields}
\usage{
define_fields(year, stat_name, substat_name, parameter, region_id, page_nr,
  ipp, nuts_nr, lau_nr, parent_chr, all_regions)
}
\arguments{
\item{year}{description}

\item{stat_name}{description}

\item{substat_name}{description}

\item{parameter}{description}

\item{region_id}{description}

\item{page_nr}{description}

\item{ipp}{description}

\item{nuts_nr}{description}

\item{lau_nr}{description}

\item{parent_chr}{description}

\item{all_regions}{all_regions}
}
\value{
A list
}
\description{
Converts given variables into field structure which are processed in query_builder().
}
