% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{dg_call}
\alias{dg_call}
\title{Make a call to the Datenguide GraphQL API}
\usage{
dg_call(region_id, years)
}
\arguments{
\item{region_id}{provide a valid region_id}

\item{years}{provide the years you want to make your request for}
}
\description{
Thins function will work with non-complex GraphQL API requests
}
