% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dg_descriptions}
\alias{dg_descriptions}
\title{Data on all available statistics and a short description}
\format{A data frame with 473 obervations and 2 variables.}
\usage{
dg_descriptions
}
\description{
A dataset containing the name and shortened description of all available
statistics.
}
\keyword{datasets}
