% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_map.R
\name{dg_map}
\alias{dg_map}
\title{dg_map}
\usage{
dg_map(lookup = "GADM3", b_color = "white", nuts_nr = 1,
  stat_name = NULL, substat_name = NULL, year = NULL,
  parameter = NULL, return_data = FALSE)
}
\arguments{
\item{lookup}{The method used to find the shapefile to use as the electoral background. Defaults to GADM}

\item{b_color}{Specify the border color}

\item{nuts_nr}{The NUTS level for the country shapefiles to download from GADM. Requires GADM to be the lookup method}

\item{stat_name}{Character string containing the name of the main statistic. Please see dg_descriptions for a full list.}

\item{substat_name}{Character string containing the name of the sub-statistic. Please see dg_descriptions for a full list. Defaults to all available sub-statistics if not specified.}

\item{year}{Numeric year(s) for which you want to retrieve the data.}

\item{parameter}{Character string containing the name(s) of the parameter(s) you want to retrieve. Please see dg_descriptions for a full list. Defaults to all available parameters if not specified.}

\item{return_data}{Return merged data instead of plot}
}
\description{
Produce a map filled with variables queried by \code{dg_call}
}
\examples{
\dontrun{
turnout_map <- dg_map(
  nuts_nr = 1,
  year = 2017,
  stat_name = "AI0506"
)
}
}
