% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_query_builder.R
\name{dg_query_builder}
\alias{dg_query_builder}
\title{Build a query in \code{GraphQL} syntax}
\usage{
dg_query_builder(field, substat_name, stat_name)
}
\arguments{
\item{field}{Field of "highest order" where the recursive function starts.
Usually query_list, defined in define fields.}

\item{substat_name}{name of the desired \code{substat}.}
}
\value{
Query for the GraphQL API
}
\description{
Builds a query in \code{GraphQL} syntax through a recursive function
that iterates over the fields defined in \code{define_fields()}. It checks for
possible arguments and subfields and pastes names and values, respectively.
}
\examples{
# dg_query_builder(field, lastfield_name)
}
