% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dg_regions}
\alias{dg_regions}
\title{Data on all available regions}
\format{A data frame with 14,677 observations and 4 variables.}
\usage{
dg_regions
}
\description{
A dataset containing the id, name, and level of all available regions.
}
\keyword{datasets}
