% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_descriptions.R
\name{dg_search}
\alias{dg_search}
\title{dg_search}
\usage{
dg_search(string)
}
\arguments{
\item{string}{String that you want to search for (accepts regex)}
}
\value{
Data frame containing meta data only containing the search string
}
\description{
Search for a string in dg_descriptions
}
\examples{
dg_descriptions <- dg_search("vote")
dg_descriptions

}
