% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_descriptions.R
\name{get_description}
\alias{get_description}
\title{dg_descriptions.R}
\usage{
get_description()
}
\value{
Data frame containing information on all statistics
}
\description{
Gets information on all statistics and their descriptions.
}
\examples{
dg_descriptions <- get_description()
dg_descriptions

}
