% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_descriptions.R
\name{get_descriptions}
\alias{get_descriptions}
\title{dg_descriptions.R}
\usage{
get_descriptions()
}
\value{
Data frame containing all available meta data
}
\description{
Gets all available meta data on statistics, sub-statistics, and parameters.
}
\examples{
dg_descriptions <- get_descriptions()
dg_descriptions

}
