% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_meta.R
\name{get_meta}
\alias{get_meta}
\title{dg_meta.R}
\usage{
get_meta()
}
\value{
Data frame containing all available meta data
}
\description{
Gets all available meta data on statistics, sub-statistics, and parameters.
}
\examples{
dg_meta <- get_meta()
dg_meta

}
