% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_call.R
\name{get_results}
\alias{get_results}
\title{dg_call.R}
\usage{
get_results(field, substat_name)
}
\arguments{
\item{field}{description}

\item{stubstat_name}{description
TODO: Update and add parameters!}
}
\value{
Data frame containing the requested data
}
\description{
Makes a call to the Datenguide GraphQL API and returns the results.
}
\examples{
dg_call(region_id = "11", year = c(2001, 2007), substat_name = 'TIERA8', parameter = c("TIERART2", "TIERART3"))

}
