% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_query_builder.R
\name{insert_page_nr}
\alias{insert_page_nr}
\title{insert_page_nr()}
\usage{
insert_page_nr(field)
}
\arguments{
\item{field}{Field to be checked if of type RegionsResult (allRegions specific type)}
}
\description{
Insert 'page, total, itemsPerPage,' after allRegions to to
ensure that those variables are always returned for allRegions queries.
}
