% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{insert_regname}
\alias{insert_regname}
\title{insert_regname()}
\usage{
insert_regname(field)
}
\arguments{
\item{field}{description}
}
\description{
Insert 'id, name,' after region to ensure that region id and name are always returned.
}
