% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_descriptions.R
\name{join_en_translation}
\alias{join_en_translation}
\title{join_en_translation}
\usage{
join_en_translation(dg_descriptions, en_list)
}
\arguments{
\item{dg_descriptions}{data set which includes all meta data}

\item{en_list}{list that includes all English translations}
}
\value{
Data frame containing all available meta data \emph{with} English translation
}
\description{
This function joins a list of English translations to the dg_descriptions dataset.
}
\examples{
dg_descriptions <- join_en_translation(datenguideR::dg_descriptions, en_list)
dg_descriptions

}
