% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_query_builder.R
\name{paste_nv}
\alias{paste_nv}
\title{paste_nv()}
\usage{
paste_nv(field)
}
\arguments{
\item{field}{The field whos name and value should be pasted into the query}
}
\description{
Pastes name(s) and value(s) of fields, used for field arguments in query_builder.
}
