% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_query_builder.R
\name{query_builder}
\alias{query_builder}
\title{Recursive query builder}
\usage{
query_builder(field, substat_name)
}
\arguments{
\item{field}{Field of "highest order" where the recursive function starts.
Usually query_list, defined in define fields.}

\item{substat_name}{name of the desired \code{substat}.}
}
\description{
Recursive query builder
}
