% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_querybuilder.R
\name{query_builder_pre}
\alias{query_builder_pre}
\title{dg_querybuilder.R}
\usage{
query_builder_pre(field, substat_name)
}
\arguments{
\item{field}{description}

\item{substat_name}{description}
}
\value{
Query builder
}
\description{
Builds the query with a recursive function iterating over fields
}
\examples{
query_builder <- function(field, substat_name)

}
