% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sum2}
\alias{sum2}
\title{Sum of vector elements.}
\usage{
sum2(..., na.rm = TRUE)
}
\arguments{
\item{...}{Numeric, complex, or logical vectors.}

\item{na.rm}{A logical scalar. Should missing values (including NaN)
be removed?}
}
\value{
If all inputs are integer and logical, then the output
will be an integer. If integer overflow
\url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
will be NA with a warning. Otherwise it will be a length-one numeric or
complex vector.

Zero-length vectors have sum 0 by definition. See
\url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{sum2} returns the sum of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
sum2(1:10)
sum2(1:5, 6:10)
sum2(F, F, F, T, T)

sum2(.Machine$integer.max, 1L)
sum2(.Machine$integer.max, 1)

\dontrun{
sum2("a")
}
}
