% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_call_results.R
\name{add_substat_info}
\alias{add_substat_info}
\title{add_substat_info}
\usage{
add_substat_info(
  api_results,
  stat_name,
  substat_name,
  parameter,
  full_descriptions,
  all_regions,
  long_format
)
}
\arguments{
\item{api_results}{description}

\item{stat_name}{description}

\item{substat_name}{description}

\item{parameter}{description}

\item{full_descriptions}{description}

\item{all_regions}{description#'}

\item{long_format}{description}
}
\value{
A tibble
}
\description{
Add Substat Info
}
