% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_call.R
\name{dg_call}
\alias{dg_call}
\title{dg_call}
\usage{
dg_call(region_id = NULL, stat_name = NULL, year = NULL,
  substat_name = NULL, parameter = NULL, ipp = NULL,
  nuts_nr = NULL, lau_nr = NULL, parent_chr = NULL,
  full_descriptions = FALSE, page_nr = NULL)
}
\arguments{
\item{region_id}{description}

\item{stat_name}{description}

\item{year}{description}

\item{substat_name}{description}

\item{parameter}{description}

\item{ipp}{description}

\item{nuts_nr}{description}

\item{lau_nr}{description}

\item{parent_chr}{description}

\item{full_descriptions}{description}

\item{page_nr}{description}
}
\value{
Data frame containing the requested data
}
\description{
Makes a call to the Datenguide GraphQL API and returns the results.
}
\examples{
dg_call(region_id = "11", 
   year = c(2001, 2007), 
   stat_name = "BETR08",
   substat_name = 'TIERA8', 
   parameter = c("TIERART2", "TIERART3"))

}
