% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_call.R
\name{dg_call}
\alias{dg_call}
\title{dg_call()}
\usage{
dg_call(region_id = NULL, stat_name = NULL, year = NULL,
  substat_name = NULL, parameter = NULL, ipp = 100, nuts_nr = NULL,
  lau_nr = NULL, parent_chr = NULL, full_descriptions = FALSE,
  page_nr = NULL)
}
\arguments{
\item{region_id}{TODO: ADD_FORMAT containing the ID of a specific region.}

\item{stat_name}{Character string containing the name of the main statistic. Please see dg_descriptions for a full list.}

\item{year}{Numeric year(s) for which you want to retrieve the data.}

\item{substat_name}{Character string containing the name of the sub-statistic. Please see dg_descriptions for a full list. Defaults to all available sub-statistics if not specified.}

\item{parameter}{Character string containing the name(s) of the parameter(s) you want to retrieve. Please see dg_descriptions for a full list. Defaults to all available parameters if not specified.}

\item{ipp}{description}

\item{nuts_nr}{TODO: ADD_FORMAT containing the number of the NUTS level. 1 refers to NUTS-1, 2 to NUTS-2, and 3 to NUTS-3.}

\item{lau_nr}{TODO: ADD_FORMAT containing the number of the LAU level.
TODO: Change this as it can currently only be "1".}

\item{parent_chr}{description}

\item{full_descriptions}{If \code{TRUE}, the returning data frame will contain the full descriptions of the
statistics as provided by GENESIS. Defaults to \code{FALSE}.}

\item{page_nr}{description}
}
\value{
Data frame containing the requested data
}
\description{
Makes a call to the Datenguide GraphQL API and returns the results.
}
\examples{
dg_call(region_id = "11", 
   year = c(2001, 2007), 
   stat_name = "BETR08",
   substat_name = 'TIERA8', 
   parameter = c("TIERART2", "TIERART3"))

}
