% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dg_descriptions}
\alias{dg_descriptions}
\title{Data on all available statistics, substatistics, and parameters}
\format{A data frame with 3,419 obervations and 7 variables.}
\usage{
dg_descriptions
}
\description{
A dataset containing the name, description, and source of all available
statistics, substatistics, and parameters.
}
\keyword{datasets}
