% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_regions.R
\name{get_region}
\alias{get_region}
\title{dg_regions.R}
\usage{
get_region(var = NULL, val = NULL)
}
\arguments{
\item{var}{Character string naming variable for filtering. Can be either id, name, level or parent.}

\item{val}{Character string naming the value for filtering.}
}
\value{
Data frame containing information on all regions
}
\description{
Imports information on all regions.
}
\examples{
dg_regions <- get_region(var = level, val = "nuts1")
dg_regions

}
