% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_regions.R
\name{get_region}
\alias{get_region}
\title{dg_regions.R}
\usage{
get_region(var = NULL, val = NULL)
}
\value{
Data frame containing information on all regions
}
\description{
Imports information on all regions.
}
\examples{
dg_regions <- get_region()
dg_regions

}
