% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_regions.R
\name{get_region}
\alias{get_region}
\title{dg_regions.R}
\usage{
get_region(var = NULL, val = NULL)
}
\arguments{
\item{var}{Character string naming the variable for filtering.
Can be either id, name, level or parent.}

\item{val}{Character string naming the value for filtering.}
}
\value{
Data frame containing geographical information on all regions
}
\description{
Imports geographical information on all regions. The data set contains regional IDs on
the NUTS-1, NUTS-2, NUTS-3 and LAU level.
}
\examples{
dg_regions <- get_region(var = level, val = "nuts1")
dg_regions

}
