% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_call.R
\name{get_results}
\alias{get_results}
\title{get_results()}
\usage{
get_results(field, stat_name, substat_name)
}
\arguments{
\item{field}{description}

\item{stat_name}{description}

\item{substat_name}{description}
}
\value{
Data frame containing the requested data
}
\description{
POST file to the GraphQL API.
}
