% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{insert_page_nr}
\alias{insert_page_nr}
\title{insert_page_nr()}
\usage{
insert_page_nr(field)
}
\arguments{
\item{field}{description}
}
\description{
Insert 'page, total, itemsPerPage,' after allRegions by default.
}
