% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_query_builder.R
\name{insert_regname}
\alias{insert_regname}
\title{insert_regname()}
\usage{
insert_regname(field)
}
\arguments{
\item{field}{Field to be checked if of type Region.}
}
\description{
If given field is of type Region 'id, name,' is inserted in query to
ensure that region id and name are always returned.
}
