% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_query_builder.R
\name{query_builder}
\alias{query_builder}
\title{dg_query_builder.R}
\usage{
query_builder(field, substat_name)
}
\arguments{
\item{field}{description}

\item{substat_name}{description}
}
\value{
Query
}
\description{
Builds the query with a recursive function iterating over fields.
}
\examples{
dg_query_builder <- function(field, substat_name)

}
